# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Akure Airport Staff Cooperative Society API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /members:
    get:
      operationId: listMembers
      tags: [members]
      summary: List all members (Admin+)
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, active, inactive]
        - in: query
          name: search
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Member"
    post:
      operationId: createMember
      tags: [members]
      summary: Create a member account (Admin+)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMemberBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
  /members/{id}:
    get:
      operationId: getMember
      tags: [members]
      summary: Get a member by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
        "404":
          description: Not found
    patch:
      operationId: updateMember
      tags: [members]
      summary: Update member (Admin+)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMemberBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
  /members/{id}/activate:
    post:
      operationId: activateMember
      tags: [members]
      summary: Activate a pending member (Admin+)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
  /members/{id}/deactivate:
    post:
      operationId: deactivateMember
      tags: [members]
      summary: Deactivate a member (Admin+)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
  /members/{id}/summary:
    get:
      operationId: getMemberSummary
      tags: [members]
      summary: Get member financial summary
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemberSummary"

  # ── SAVINGS ──────────────────────────────────────────────────────────────────
...
  /savings/{memberId}:
    get:
      operationId: getMemberSavings
      tags: [savings]
      summary: Get savings balance for a member (Admin+)
      parameters:
        - in: path
          name: memberId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavingsBalance"

  # ── TRANSACTIONS ─────────────────────────────────────────────────────────────
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List current user's notifications
      parameters:
        - in: query
          name: unread
          schema:
            type: boolean
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
  /notifications/{id}/read:
    post:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a notification as read
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
components:
  schemas:
...
    Member:
      type: object
      properties:
        id:
          type: integer
        clerkUserId:
          type: ["string", "null"]
        fullName:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        staffId:
          type: ["string", "null"]
        role:
          type: string
          enum: [member, admin, financial_auditor, treasurer, super_admin]
        status:
          type: string
          enum: [pending, active, inactive]
        savingsBalance:
          type: number
        totalLoanBalance:
          type: number
        totalStoreDebt:
          type: number
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - fullName
        - email
        - role
        - status
        - savingsBalance
        - totalLoanBalance
        - totalStoreDebt
        - createdAt
        - updatedAt
...
    CreateMemberBody:
      type: object
      properties:
        fullName:
          type: string
        email:
          type: string
        phone:
          type: string
        staffId:
          type: string
        role:
          type: string
          enum: [member, admin, financial_auditor, treasurer, super_admin]
        status:
          type: string
          enum: [pending, active, inactive]
      required: [fullName, email]
    UpdateMemberBody:
      type: object
      properties:
        fullName:
          type: string
        phone:
          type: string
        staffId:
          type: string
        role:
          type: string
          enum: [member, admin, financial_auditor, treasurer, super_admin]
        status:
          type: string
          enum: [pending, active, inactive]
    MemberSummary:
      type: object
      properties:
        memberId:
          type: integer
        fullName:
          type: string
        savingsBalance:
          type: number
        totalLoanBalance:
          type: number
        totalStoreDebt:
          type: number
        totalSavingsContributed:
          type: number
        totalLoansRepaid:
          type: number
        activeLoansCount:
          type: integer
        pendingStoreDebt:
          type: number
      required:
        - memberId
        - fullName
        - savingsBalance
        - totalLoanBalance
        - totalStoreDebt
        - totalSavingsContributed
        - totalLoansRepaid
        - activeLoansCount
        - pendingStoreDebt

    # ── SAVINGS ─────────────────────────────────────────────────────────────────
    SavingsBalance:
      type: object
      properties:
        memberId:
          type: integer
        balance:
          type: number
        lastUpdated:
          type: string
          format: date-time
      required: [memberId, balance, lastUpdated]

    # ── TRANSACTIONS ─────────────────────────────────────────────────────────────
...
    Notification:
      type: object
      properties:
        id:
          type: integer
        memberId:
          type: integer
        type:
          type: string
          enum: [loan_update, transaction, store_purchase, system]
        title:
          type: string
        message:
          type: string
        isRead:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, memberId, type, title, message, isRead, createdAt]

    # ── AUDIT LOG ──────────────────────────────────────────────────────────────
...